requirements.txt
-i https://pypi.org/simple
absl-py==1.3.0 ; python_version >= '3.6'
astunparse==1.6.3
attrs==22.1.0 ; python_version >= '3.5'
beautifulsoup4==4.11.1 ; python_full_version >= '3.6.0'
cachetools==5.2.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
contourpy==1.0.6 ; python_version >= '3.7'
cycler==0.11.0 ; python_version >= '3.6'
deepface==0.0.75
filelock==3.8.2 ; python_version >= '3.7'
fire==0.5.0
flask==2.2.2 ; python_version >= '3.7'
flatbuffers==22.12.6
fonttools==4.38.0 ; python_version >= '3.7'
gast==0.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gdown==4.6.0
google-auth==2.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-oauthlib==0.4.6 ; python_version >= '3.6'
google-pasta==0.2.0
grpcio==1.51.1 ; python_version >= '3.7'
h5py==3.7.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
git+https://github.com/ocampor/image-quality@a0435e2966883b3420ee8be16a3ef9c6a8de79aa#egg=image-quality
imageio==2.23.0 ; python_version >= '3.7'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
keras==2.11.0 ; python_version >= '3.7'
kiwisolver==1.4.4 ; python_version >= '3.7'
libclang==14.0.6
libsvm==3.23.0.4 ; python_version >= '3.5'
markdown==3.4.1 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
matplotlib==3.6.2 ; python_version >= '3.8'
mediapipe==0.9.0.1
mtcnn==0.1.1
networkx==2.8.8 ; python_version >= '3.8'
numpy==1.24.0 ; python_version >= '3.8'
oauthlib==3.2.2 ; python_version >= '3.6'
opencv-contrib-python==4.6.0.66 ; python_version >= '3.6'
opencv-python==4.6.0.66 ; python_version >= '3.6'
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0 ; python_version >= '3.5'
packaging==22.0 ; python_version >= '3.7'
pandas==1.5.2 ; python_version >= '3.8'
pillow==9.3.0
protobuf==3.19.6 ; python_version >= '3.5'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
pysocks==1.7.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7
pywavelets==1.4.1 ; python_version >= '3.8'
requests[socks]==2.28.1 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
retina-face==0.0.12 ; python_full_version >= '3.5.5'
rsa==4.9 ; python_version >= '3.6'
scikit-image==0.19.3 ; python_version >= '3.7'
scipy==1.9.3 ; python_version >= '3.8'
setuptools==65.6.3 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1 ; python_version >= '3.6'
tensorboard==2.11.0 ; python_version >= '3.7'
tensorboard-data-server==0.6.1 ; python_version >= '3.6'
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0 ; python_version >= '3.7'
tensorflow-estimator==2.11.0 ; python_version >= '3.7'
tensorflow-io-gcs-filesystem==0.29.0 ; platform_machine != 'arm64' or platform_system != 'Darwin'
termcolor==2.1.1 ; python_version >= '3.7'
tifffile==2022.10.10 ; python_version >= '3.8'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.2 ; python_version >= '3.7'
wheel==0.38.4 ; python_version >= '3.7'
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wsq==0.5
